requirements.txt
streamlit
streamlit_option_menu
torch
torchvision
timm
pillow
